+++ r-queues.tf
@@ -1,5 +1,5 @@
 resource "azurerm_storage_queue" "queue" {
-  for_each = { for q in var.queues : q.name => q }
+  for_each = try({ for q in var.queues : q.name => q }, {})
 
   storage_account_name = azurerm_storage_account.storage.name