+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Fixed
   * [GH-2](https://github.com/claranet/terraform-azurerm-storage-account/pull/2): Enable logging only if queues are used/created
 
+Changed
+  * AZ-130: Module code improvements, bump `diagnostics` module for sub-entities
+
 # v7.1.0 - 2022-10-21
 
 Added