+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v7.2.0 - 2022-11-10
 
 Fixed
   * [GH-2](https://github.com/claranet/terraform-azurerm-storage-account/pull/2): Enable logging only if queues are used/created